requirements.txt
aiohttp==3.11.18
pandas==2.2.3
Pillow==11.2.1
python-dotenv==1.1.0
redis==5.2.1
